// src/interfaces/Form.ts
export interface FormFieldsProps {
  id?: string
  firstName: string
  firstSurname: string
  secondSurname: string
  lastName: string
  office: string
  birthdate: string
  email: string
  identificationNumber: string
  identificationType: string
  phone: string
  jornal: string
  comments: string
  expeditionDateDocument: string
}